refactor(video-conf): move the call window's components under components/
`CallBar/` and `CallPanel/` already lived in `views/conference/components/`,
while the rest of the call window's presentational pieces sat loose beside the
views they are used by — so in Storybook they appeared as siblings of the
`components` folder rather than inside it.

Moved, one folder per component, which is what `CallPanel/` already did for a
single component. `CallPanelHeader` and `ConferenceIframe` go too: both are
presentational, take everything they show as props, and own no routing or data.

`ConferencePreflight` stays a view — it is a screen, with its own state and its
own decisions — as do the `Conference*Page`/`Route`/`Room`/`Thread`/`Chat`
files. `panelStyles.ts` and `testFixtures.ts` stay where they are, shared by
both sides.

Nothing changes but import paths.
