+++ apps/meteor/client/views/conference/components/CallMemberItem/CallMemberItem.tsx
@@ -4,12 +4,12 @@ import { UserAvatar } from '@rocket.chat/ui-avatar';
 import { useSetting } from '@rocket.chat/ui-contexts';
 import { useTranslation } from 'react-i18next';
 
-import type { ConferenceMember } from './hooks/useConferenceEmbedded';
-import { getUserDisplayNames } from '../../../lib/getUserDisplayNames';
-import type { ConferenceMemberStatus } from '../../../lib/videoConference/memberStatus';
-import { canRingConferenceMember, getConferenceMemberStatus } from '../../../lib/videoConference/memberStatus';
-import { ReactiveUserStatus } from '../../components/UserStatus';
-import { useRingingExpiry } from '../../hooks/useRingingExpiry';
+import { getUserDisplayNames } from '../../../../../lib/getUserDisplayNames';
+import type { ConferenceMemberStatus } from '../../../../../lib/videoConference/memberStatus';
+import { canRingConferenceMember, getConferenceMemberStatus } from '../../../../../lib/videoConference/memberStatus';
+import { ReactiveUserStatus } from '../../../../components/UserStatus';
+import { useRingingExpiry } from '../../../../hooks/useRingingExpiry';
+import type { ConferenceMember } from '../../hooks/useConferenceEmbedded';
 
 type CallMemberItemProps = {
 	member: ConferenceMember;
