+++ apps/meteor/client/views/conference/components/CallPanelHeader/CallPanelHeader.tsx
